task2 modules

+++ sample/db.py
@@ -1,6 +1,7 @@
 import click
 import os
 import sqlite3
+import sys
 
 
 def get_full_path_to_db(dbname):
@@ -9,21 +10,19 @@ def get_full_path_to_db(dbname):
 
 
 @click.command()
+@click.option('--dbname', prompt='Enter database name')
 @click.option('--username', prompt='Enter a username to be added')
-def add_user(username):
-    dbname = 'pam'
+def add_user(dbname, username):
     path_to_db = get_full_path_to_db(dbname)
-    # if os.path.exists(path_to_db):
-    #     print('Database {}.db already exists'.format(dbname))
-    #     return
+    if not os.path.exists(path_to_db):
+        print('Database {}.db does not exist, can\'t update'.format(dbname))
+        return
     conn = sqlite3.connect(path_to_db)
     cursor = conn.cursor()
-    # select_max = cursor.execute("SELECT MAX(userid) FROM table1").fetchall()[0][0]
-    # max_id = int(select_max) if select_max is not None else 0
-    # print(max_id)
     cursor.execute("INSERT INTO table1 VALUES (NULL, '{}')".format(username))
     conn.commit()
-    print(cursor.execute('SELECT * FROM table1').fetchall())
+    print('Added username "{}" into {}.db'.format(username, dbname))
+    # print(cursor.execute('SELECT * FROM table1').fetchall())
 
 
 @click.command()
@@ -32,13 +31,16 @@ def add_user(username):
 def delete_user(dbname, username):
     path_to_db = get_full_path_to_db(dbname)
     if not os.path.exists(path_to_db):
-        print('Database {}.db does not exist'.format(dbname))
+        print('Database {}.db does not exist, nothing to delete'.format(dbname))
         return
     conn = sqlite3.connect(path_to_db)
     cursor = conn.cursor()
-    cursor.execute("DELETE FROM table1 WHERE username = '{}'".format(username))
-    conn.commit()
-    print(cursor.execute('SELECT * FROM table1').fetchall())
+    if cursor.execute('SELECT * FROM table1 WHERE username="{}"'.format(username)).fetchone():
+        cursor.execute("DELETE FROM table1 WHERE username = '{}'".format(username))
+        conn.commit()
+        print('Deleted username "{}" from {}.db'.format(username, dbname))
+    else:
+        print('No username {} in database {}, nothing to delete'.format(username, dbname))
 
 
 @click.command()
@@ -56,18 +58,21 @@ def create_table(dbname):
 
 
 @click.command()
-@click.option('--action', prompt='Choose an action to perform (CUD)', type=click.Choice(['C', 'U', 'D']))
+@click.option('--action', prompt='Choose an action to perform (CUD, or Q for quit)', type=click.Choice(['C', 'U', 'D', 'Q']))
 def do_stuff(action):
     if action == 'C':
-        create_table()
+        create_table(standalone_mode=False)
     elif action == 'U':
-        add_user()
+        add_user(standalone_mode=False)
     elif action == 'D':
-        delete_user()
+        delete_user(standalone_mode=False)
+    elif action == 'Q':
+        sys.exit(0)
 
 
 def main():
-    do_stuff()
+    while True:
+        do_stuff(standalone_mode=False)
 
 
 if __name__ == '__main__':
